refactor: 리뷰 별점 데이터 타입 변경
- Double 타입으로 변경하여 소숫점 별점에도 대응하도록 변경
- 리뷰 저장 로직 일부 수정

+++ src/main/java/com/dongyang/dongpo/domain/store/StoreReview.java
@@ -28,7 +28,7 @@ public class StoreReview {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    private Integer reviewStar;
+    private Double reviewStar;
 
     @Column(columnDefinition = "TEXT")
     private String text;
